refactor: move agent builder to test utils
