+++ crates/core/src/factories/mem_peer_store.rs
@@ -256,3 +256,7 @@ impl Inner {
 
 #[cfg(test)]
 mod test;
+
+/// Peer store related utilities for testing.
+#[cfg(test)]
+pub mod test_utils;
